+++ src/main/java/com/heroiclabs/nakama/DefaultClient.java
@@ -210,22 +210,22 @@ public class DefaultClient implements Client {
 
     @Override
     public SocketClient createSocket(final int port, final int socketTimeoutMs) {
-        return new WebSocketClient(host, port, ssl, socketTimeoutMs, WebSocketClient.DEFAULT_PING_MS, this.trace, Executors.newSingleThreadExecutor());
+        return new WebSocketClient(host, port, ssl, socketTimeoutMs, WebSocketClient.DEFAULT_PING_MS, this.trace, null);
     }
 
     @Override
     public SocketClient createSocket(final String host, final int port, final boolean ssl) {
-        return new WebSocketClient(host, port, ssl, WebSocketClient.DEFAULT_TIMEOUT_MS, WebSocketClient.DEFAULT_PING_MS, this.trace, Executors.newSingleThreadExecutor());
+        return new WebSocketClient(host, port, ssl, WebSocketClient.DEFAULT_TIMEOUT_MS, WebSocketClient.DEFAULT_PING_MS, this.trace, null);
     }
 
     @Override
     public SocketClient createSocket(final String host, final int port, final boolean ssl, final int socketTimeoutMs) {
-        return new WebSocketClient(host, port, ssl, socketTimeoutMs, WebSocketClient.DEFAULT_PING_MS, this.trace, Executors.newSingleThreadExecutor());
+        return new WebSocketClient(host, port, ssl, socketTimeoutMs, WebSocketClient.DEFAULT_PING_MS, this.trace, null);
     }
 
     @Override
     public SocketClient createSocket(final String host, final int port, final boolean ssl, final int socketTimeoutMs, final int socketPingMs) {
-        return new WebSocketClient(host, port, ssl, socketTimeoutMs, socketPingMs, this.trace, Executors.newSingleThreadExecutor());
+        return new WebSocketClient(host, port, ssl, socketTimeoutMs, socketPingMs, this.trace, null);
     }
 
     @Override
@@ -1791,7 +1791,7 @@ public class DefaultClient implements Client {
 
     @Override
     public ListenableFuture<Empty> logoutSession(@NonNull final Session session) {
-        return autoRefreshSession(session, getStub(session).sessionLogout(SessionLogoutRequest.newBuilder().setToken(session.getAuthToken()).setRefreshToken(session.getRefreshToken()).build()));
+        return getStub(session).sessionLogout(SessionLogoutRequest.newBuilder().setToken(session.getAuthToken()).setRefreshToken(session.getRefreshToken()).build());
     }
 
     @Override